// loans/frontend/src/app/api/model/contract-dto.ts
/**
 * traefik-microstack API Gateway
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AssetDto } from './asset-dto';


/**
 * Contract data
 */
export interface ContractDto { 
    /**
     * Client identifier
     */
    client?: string;
    /**
     * Contract code/number
     */
    code?: string;
    /**
     * Application number - business identifier for the application
     */
    applicationNumber?: number;
    /**
     * Version identifier for the application number
     */
    applicationNumberVersion?: number;
    /**
     * Contract name
     */
    name?: string;
    /**
     * Contract start date
     */
    startDate?: string;
    /**
     * Contract end date
     */
    endDate?: string;
    /**
     * Opening amount of the contract
     */
    openingAmount?: number;
    /**
     * Automatic invoicing residual value flag
     */
    automaticInvoicingRv?: boolean;
    /**
     * Contract duration in years and months
     */
    duration?: string;
    /**
     * Negotiated interest rate
     */
    negotiatedRate?: number;
    /**
     * Calculated interest rate
     */
    calculatedRate?: number;
    /**
     * Estimated start date of the contract
     */
    estimatedStartDate?: string;
    /**
     * Contract creation user
     */
    createUser?: string;
    /**
     * Contract creation date
     */
    createDate?: string;
    /**
     * Last change user
     */
    lastChangeUser?: string;
    /**
     * Last change date
     */
    lastChangeDate?: string;
    /**
     * Contract approval date
     */
    approvalDate?: string;
    /**
     * Contract signing date
     */
    signedDate?: string;
    /**
     * Payment date
     */
    paymentDate?: string;
    /**
     * Down payment amount
     */
    downPayment?: number;
    /**
     * Salesperson information
     */
    salesperson?: string;
    /**
     * Condition key
     */
    conditionKey?: string;
    /**
     * Condition text
     */
    conditionText?: string;
    /**
     * Billing person information
     */
    billingPerson?: string;
    /**
     * Billing text
     */
    billingText?: string;
    /**
     * Contentious flag
     */
    contentious?: boolean;
    /**
     * Branding identifier
     */
    branding?: string;
    /**
     * Business channel identifier
     */
    businessChannel?: string;
    /**
     * Business type identifier
     */
    businessType?: string;
    /**
     * Contract status
     */
    status?: string;
    /**
     * Currency identifier
     */
    currency?: string;
    /**
     * Department identifier
     */
    department?: string;
    /**
     * Distribution channel identifier
     */
    distributionChannel?: string;
    /**
     * Financing type identifier
     */
    financingType?: string;
    /**
     * Product identifier
     */
    product?: string;
    /**
     * Rate type identifier
     */
    rateType?: string;
    /**
     * List of assets associated with this contract
     */
    assets?: Array<AssetDto>;
}