// loans/frontend/src/app/api/model/contract-search-dto.ts
/**
 * traefik-microstack API Gateway
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Contract search result containing comprehensive contract and related information
 */
export interface ContractSearchDto { 
    /**
     * Contract number - primary contract identifier
     */
    contractNumber?: string;
    /**
     * Application/Request number
     */
    applicationNumber?: number;
    /**
     * Partner number - primary partner identifier
     */
    partnerNumber?: number;
    /**
     * First name of the contract holder
     */
    firstName?: string;
    /**
     * Last name or company name of the contract holder
     */
    name?: string;
    /**
     * Date of birth of the contract holder
     */
    dateOfBirth?: string;
    /**
     * Postal code of the contract holder\'s address
     */
    postalCode?: string;
    /**
     * City of the contract holder\'s address
     */
    city?: string;
    /**
     * Vehicle license plate number
     */
    vehicleLicensePlate?: string;
    /**
     * Street name of the contract holder\'s address
     */
    street?: string;
    /**
     * House number of the contract holder\'s address
     */
    houseNumber?: string;
    /**
     * Client identifier for multi-tenant scenarios
     */
    client?: string;
}